Fixes for upload driver
When `$value` was not always trimmed it caused a server 500 error on import.  The error log was showing a line break e.g.

```
Fatal Error: ErrorException 0 - filesize(): stat failed for /path-redacted/workspace/images/
filename.jpg on line 39 of /path-redacted/extensions/importcsv/drivers/ImportDriver_upload.php
```

Also, the resulting import set a path for `$fileData['file']` because `$filename` was used I believe, so Symphony was complaining the file did not exist even though it was there and the link worked.  Undoing the import, changing `$fileData['file']` to take the trimmed `$value` made the import work afterwards.